fix(linux): cast thing enums in fmt exception paths

+++ Tibia/silnik/canary_test/testyy/src/client/thingtype.cpp
@@ -480,7 +480,11 @@ void ThingType::unserialize(const uint16_t clientId, const ThingCategory categor
     }
 
     if (!done)
-        throw Exception("corrupt data (id: {}, category: {}, count: {}, lastAttr: {})", m_id, m_category, count, attr);
+        throw Exception("corrupt data (id: {}, category: {}, count: {}, lastAttr: {})",
+            m_id,
+            static_cast<uint8_t>(m_category),
+            count,
+            static_cast<uint8_t>(attr));
 
     const bool hasFrameGroups = category == ThingCategoryCreature && g_game.getFeature(Otc::GameIdleAnimations);
     const uint8_t groupCount = hasFrameGroups ? fin->getU8() : 1;
